fix: money type
입력한 돈이 int형이라 큰 정수를 처리하지 못하는 경우 수정
- int형에서 long형으로 변경하였음

+++ src/main/java/lotto/controller/MainController.java
@@ -25,7 +25,7 @@ public class MainController {
         OutputView.printBenefitRatio(result.getBenefitRatio(money.getMoney()));
     }
 
-    private static List<Lotto> makeLotto(int count) {
+    private static List<Lotto> makeLotto(long count) {
         List<List<Integer>> lottoNumList = generateNum(count);
         List<Lotto> lottoList = new ArrayList<>();
 
@@ -41,7 +41,7 @@ public class MainController {
         return winLotto;
     }
 
-    private static List<List<Integer>> generateNum(int count) {
+    private static List<List<Integer>> generateNum(long count) {
         List<List<Integer>> lottoNumList = new ArrayList<>();
 
         while (lottoNumList.size() < count) {
